Separate controllers into their own files, remove empty controllers, add $table_name to models
