+++ src/PageTypes/EventPageController.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace CWP\CWP\PageTypes;
+
+use PageController;
+
+class EventPageController extends PageController
+{
+
+}
